LibC: Implement {f,}getwc()
